Bias in Media
Our group created a study where half the participants were shown a treatment video that in-depth described bias and how media frames and primes it. The goal of the study was to see if participants were able to identify valence words in news article titles and identify bias within those titles.